Instantiate pre-defined circuit
